+++ CHANGELOG.md
@@ -4,13 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/) and this
 project adheres to [Semantic Versioning](https://semver.org/).
 
-## [0.1.3] - Unreleased
+## [0.2.0] - 2024-07-10
 ### Added
 - Material Design icons catalog.
 - Material Symbols catalog.
 
+### Removed
+- Some icons of SimpleIcons catalog have been removed in [13.0.0](https://github.com/simple-icons/simple-icons/releases/tag/13.0.0).
+
 ### Fixed
 - Heroicons catalog.
+- Updated dependencies.
 
 ## [0.1.2] - 2024-06-21
 ### Added
@@ -23,7 +27,7 @@ project adheres to [Semantic Versioning](https://semver.org/).
 ## [0.1.0] - 2024-06-15
 First version
 
-[0.1.3]: https://github.com/lumeland/icon-plugins/compare/v0.1.2...HEAD
+[0.2.0]: https://github.com/lumeland/icon-plugins/compare/v0.1.2...v0.2.0
 [0.1.2]: https://github.com/lumeland/icon-plugins/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/lumeland/icon-plugins/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/lumeland/icon-plugins/releases/tag/v0.1.0